🌱 Changelog: - Fixed an issue that caused platform textures to appear offset 4px to the right.
🐞 Known Bugs:
-

📋 Backlog:
-

+++ GeoWareMod/src/main/java/com/github/razorplay01/razorplayapi/util/texture/Animation.java
@@ -46,7 +46,7 @@ public class Animation {
         return textures.get(currentFrame);
     }
 
-    public int getCurrentU() {
+    public float getCurrentU() {
         return getCurrentTexture().u();
     }
